requirements.txt
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django==4.1.3
djangorestframework==3.14.0
matplotlib==3.6.2
mediapipe==0.8.11
notebook==6.5.2
opencv-python==4.5.5.62
opencv-contrib-python==4.5.5.62
pandas==1.5.1
psycopg2==2.9.5
sklearn==0.0.post1
tensorflow==2.10.0
tensorflow-gpu==2.10.0
